Moved cloud event form to runtime-tools-enveloped-components
