+++ packages/runtime-tools-enveloped-components/package.json
@@ -43,7 +43,8 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-json-view": "^1.21.3",
-    "react-moment": "0.9.7"
+    "react-moment": "0.9.7",
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
     "@babel/core": "^7.16.0",
@@ -60,6 +61,7 @@
     "@types/react": "^17.0.6",
     "@types/react-dom": "^17.0.5",
     "@types/testing-library__jest-dom": "^5.9.1",
+    "@types/uuid": "^8.3.0",
     "jest": "^26.6.3",
     "jest-junit": "^14.0.0",
     "jest-when": "^3.5.0",
