+++ packages/runtime-tools-enveloped-components/src/cloudEventForm/embedded/EmbeddedCloudEventFormChannelApiImpl.ts
