+++ packages/runtime-tools-enveloped-components/src/cloudEventForm/embedded/index.ts
