+++ packages/runtime-tools-enveloped-components/src/cloudEventForm/envelope/CloudEventFormEnvelopeViewDriver.tsx
@@ -17,9 +17,9 @@
  * under the License.
  */
 
-import { MessageBusClientApi } from "@kie-tools-core/envelope-bus/dist/api";
 import { CloudEventFormChannelApi, CloudEventFormDriver } from "../api";
 import { CloudEventRequest } from "@kie-tools/runtime-tools-gateway-api/dist/types";
+import { MessageBusClientApi } from "@kie-tools-core/envelope-bus/dist/api";
 
 export class CloudEventFormEnvelopeViewDriver implements CloudEventFormDriver {
   constructor(private readonly channelApi: MessageBusClientApi<CloudEventFormChannelApi>) {}
