+++ packages/runtime-tools-enveloped-components/src/cloudEventForm/envelope/components/CloudEventCustomHeadersEditor/CloudEventCustomHeadersEditor.tsx
@@ -25,7 +25,7 @@ import { Stack, StackItem } from "@patternfly/react-core/dist/js/layouts/Stack";
 import { TextInput } from "@patternfly/react-core/dist/js/components/TextInput";
 import PlusCircleIcon from "@patternfly/react-icons/dist/esm/icons/plus-circle-icon";
 import TrashIcon from "@patternfly/react-icons/dist/esm/icons/trash-icon";
-import { componentOuiaProps, OUIAProps } from "../../../../../ouiaTools";
+import { componentOuiaProps, OUIAProps } from "@kie-tools/runtime-tools-components/dist/ouiaTools";
 
 export interface CloudEventCustomHeadersEditorApi {
   reset(): void;
