+++ packages/runtime-tools-enveloped-components/src/cloudEventForm/envelope/components/CloudEventFieldLabelIcon/CloudEventFieldLabelIcon.tsx
